Move FunctionDefinition out to its own file
