--- layer/decorators/definitions.py
@@ -1,1 +1,0 @@
-from layer.contracts.runs import FunctionDefinition  # noqa
